+++ src/commands/add.rs
@@ -1,11 +1,12 @@
 use anyhow::Result;
 use colored::*;
 
-use crate::commands::update::run_update;
+use std::collections::HashSet;
+
+use crate::commands::update::add_package_into_project;
 use crate::emoji;
-use crate::package::PackageJson;
 use crate::registry::Registry;
-use crate::utils::{self, work_dirs};
+use crate::utils;
 
 /// Add logic
 pub fn add(
@@ -13,14 +14,24 @@ pub fn add(
     package_name_version: &str,
     is_dev: bool,
     pure: bool,
+    resolve_workspace: bool,
 ) -> Result<()> {
     let (package_name, package_version) = utils::package_name_version_parse(package_name_version);
 
     utils::validate_version_in_registry(registry, package_name, package_version);
 
-    let dirs = work_dirs(package_name)?;
+    let project_dir = std::env::current_dir()?;
+    let mut visited = HashSet::new();
 
-    run_update(registry, package_name, &std::env::current_dir()?)?;
+    add_package_into_project(
+        registry,
+        package_name,
+        is_dev,
+        pure,
+        resolve_workspace,
+        &project_dir,
+        &mut visited,
+    )?;
 
     if pure {
         println!(
@@ -29,12 +40,7 @@ pub fn add(
                 .italic()
                 .bright_black()
         );
-    } else {
-        // --- Automate package.json modification ---
-        PackageJson::update_dependency(&dirs.project_dir, package_name, is_dev)?;
     }
-
-    registry.add_package_installation(package_name, &dirs.project_dir)?;
 
     println!(
         "{}\n{}",